Navigation on startup, Pages renamed to Views for MVVM
