+++ Valcoin/Views/MiningPage.xaml
@@ -2,16 +2,19 @@
 <!-- Licensed under the MIT License. -->
 
 <Page
-    x:Class="Valcoin.Pages.MiningPage"
+    x:Class="Valcoin.Views.MiningPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Valcoin.Pages"
+    xmlns:local="using:Valcoin.Views"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
 
-    <Grid Background="Blue">
-        <TextBlock Text="Mining Page" HorizontalAlignment="Center" VerticalAlignment="Center"/>
+    <Grid>
+        <TextBlock Text="Mining Page"/>
+        <Button Content="Start"
+                Click="Button_Click"/>
+        <Button Content="Stop"/>
     </Grid>
 </Page>
