+++ Valcoin/Views/MiningPage.xaml.cs
@@ -19,7 +19,7 @@ using Windows.Foundation.Collections;
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
 
-namespace Valcoin.Pages
+namespace Valcoin.Views
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
@@ -29,6 +29,11 @@ namespace Valcoin.Pages
         public MiningPage()
         {
             this.InitializeComponent();
+        }
+
+        private void Button_Click(object sender, RoutedEventArgs e)
+        {
+            // TODO: Start a BackgroundWorker to begin the Miner, and let the miner check on each hash if the backgroundworker has a stop requested
         }
     }
 }
